backend/rag/requirements.txt
boto3==1.40.17
pymongo==4.6.0
langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.29
langchain-anthropic==0.3.13
langchain-aws==0.2.19
python-dotenv==1.0.1
Unidecode==1.3.8

